feat: admin panel

+++ app/Http/Controllers/AuthController.php
@@ -29,6 +29,10 @@ class AuthController extends Controller
             ],401);
         }
 
+        if ($user->banned) {
+            return response()->json(['message' => 'Sua conta foi suspensa'],403);
+        }
+
         $user->tokens()->delete();
         $token = $user->createToken('auth')->plainTextToken;
